+++ example/single/lib/app_images.dart
@@ -1,7 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-class AppImages {
-  const AppImages._();
-
-  static const String shieldHero = 'assets/images/shieldHero.jpg';
-}