+++ example/single/lib/r.dart
@@ -1,0 +1,11 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: library_private_types_in_public_api
+
+class R {
+  const R._();
+
+  static const String imagesShieldHero = 'assets/images/shieldHero.jpg';
+
+  static const String mainImage = 'assets/mainImage.jpg';
+}