+++ lib/mek_assets.dart
@@ -4,6 +4,6 @@
 library mek_generate_assets;
 
 export 'src/build_group/build_group_node.dart';
-export 'src/build_group/build_multi_group_node.dart';
+export 'src/build_group/build_tree_group_node.dart';
 export 'src/generate_assets.dart';
 export 'src/settings/settings.dart';