+++ lib/src/build_group/build_group_node.g.dart
@@ -11,23 +11,23 @@ part of 'build_group_node.dart';
 mixin _$Node {
   Node get _self => this as Node;
 
-  List<Object?> get _props => [
-        _self.dirName,
-        _self.filePaths,
-        _self.children,
-      ];
+  Iterable<Object?> get _props sync* {
+    yield _self.dirName;
+    yield _self.files;
+    yield _self.children;
+  }
 
-  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is _$Node &&
           runtimeType == other.runtimeType &&
-          const DeepCollectionEquality().equals(_props, other._props);
+          DataClass.$equals(_props, other._props);
 
-  @override
+  int get hashCode => Object.hashAll(_props);
+
   String toString() => (ClassToString('Node')
         ..add('dirName', _self.dirName)
-        ..add('filePaths', _self.filePaths)
+        ..add('files', _self.files)
         ..add('children', _self.children))
       .toString();
 }