+++ lib/src/generate_assets.dart
@@ -7,21 +7,26 @@ import 'package:glob/glob.dart';
 import 'package:mek_assets/src/build_group/build_group_classes.dart';
 import 'package:mek_assets/src/build_group/build_multi_group_node.dart';
 import 'package:mek_assets/src/build_group/build_single_group_node.dart';
+import 'package:mek_assets/src/build_group/build_tree_group_node.dart';
 import 'package:mek_assets/src/settings/settings.dart';
 import 'package:mek_assets/src/utils.dart';
 import 'package:path/path.dart' as pt;
-import 'package:pure_extensions/pure_extensions.dart';
 
 void generateCode(Settings settings) {
   final fileSystem = LocalFileSystem();
   final assets = <String>[];
   final buildGroup = settings.format.fold(single: () {
-    return BuildMultiGroupNode(
+    return BuildSingleGroupNode(
       fileSystem: LocalFileSystem(),
       options: settings,
     );
   }, multi: () {
-    return BuildSingleGroupNode(
+    return BuildMultiGroupNode(
+      fileSystem: LocalFileSystem(),
+      options: settings,
+    );
+  }, tree: () {
+    return BuildTreeGroupNode(
       fileSystem: LocalFileSystem(),
       options: settings,
     );
@@ -31,14 +36,12 @@ void generateCode(Settings settings) {
   );
 
   for (final groupSettings in settings.groups) {
-    final outputFileName = '${groupSettings.resolveClassName(settings).snakeCase}.dart';
-    final outputDir = groupSettings.resolveOutputDir(settings);
-    final outputFile = fileSystem.file(pt.join(outputDir, outputFileName));
+    final outputFile = fileSystem.file(groupSettings.resolveOutputFile(settings));
 
     final filePaths = groupSettings.include.expand((glob) {
       return Glob(pt.posix.join(groupSettings.inputDir, glob)).listFileSystemSync(fileSystem);
     }).where((e) {
-      return e.statSync().type == FileSystemEntityType.file;
+      return e.statSync().type == FileSystemEntityType.file && !e.basename.startsWith('.');
     }).map((file) {
       return file.path.substring(groupSettings.inputDir.length + 3); // ./<PATH>/
     }).toList();
@@ -51,14 +54,16 @@ void generateCode(Settings settings) {
       continue;
     }
 
-    final root = buildGroup(groupSettings, filePaths);
-    final classes = buildClasses(groupSettings, root);
+    final roots = buildGroup(groupSettings, filePaths);
+    final classes = roots.expand((root) => buildClasses(groupSettings, root));
 
     final library = Library((b) => b..body.addAll(classes));
 
     final emitter = DartEmitter(useNullSafetySyntax: true);
     final formatter = DartFormatter(pageWidth: settings.pageWidth);
-    final rawCode = '// GENERATED CODE - DO NOT MODIFY BY HAND\n\n${library.accept(emitter)}';
+    final rawCode = '// GENERATED CODE - DO NOT MODIFY BY HAND\n\n'
+        '// ignore_for_file: library_private_types_in_public_api\n\n'
+        '${library.accept(emitter)}';
     final code = formatter.format(rawCode);
 
     outputFile.createSync(recursive: true);