+++ lib/src/settings/settings.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:path/path.dart' as pt;
+import 'package:pure_extensions/pure_extensions.dart';
 
 part 'settings.g.dart';
 
@@ -15,7 +16,7 @@ class YamlSettings {
 }
 
 @JsonEnum()
-enum BuildFormat { single, multi }
+enum BuildFormat { single, multi, tree }
 
 @JsonSerializable()
 class Settings {
@@ -38,32 +39,48 @@ class Settings {
 class GroupSettings {
   final String inputDir;
   final List<String> include;
-  final String className;
+  final String? prefixClassName;
+  final String? className;
   final String outputDir;
+  final String? outputFileName;
 
   const GroupSettings._({
     required this.inputDir,
-    this.className = 'R',
+    this.prefixClassName,
+    this.className,
     this.include = const ['**'],
     this.outputDir = 'lib',
+    this.outputFileName,
   });
 
   factory GroupSettings({
     required String inputDir,
-    String className = 'R',
+    String? prefixClassName,
+    String? className,
     List<String> include = const ['**'],
     String outputDir = 'lib',
+    String? outputFileName,
   }) {
     return GroupSettings._(
       inputDir: pt.posix.normalize(inputDir),
+      prefixClassName: prefixClassName,
       className: className,
       include: include.isEmpty ? const ['**'] : include,
       outputDir: outputDir,
+      outputFileName: outputFileName,
     );
   }
 
-  String resolveClassName(Settings settings) => className;
-  String resolveOutputDir(Settings settings) => outputDir;
+  String resolveClassName(Settings settings, [String dir = '.']) {
+    final dirName = pt.basename(dir);
+    final partialName =
+        dirName == '.' ? (className ?? pt.basename(inputDir).pascalCase) : dirName.pascalCase;
+    return prefixClassName != null ? '$prefixClassName$partialName' : partialName;
+  }
+
+  String resolveOutputFile(Settings settings) {
+    return pt.join(outputDir, '${outputFileName ?? resolveClassName(settings).snakeCase}.dart');
+  }
 
   factory GroupSettings.fromJson(Map<dynamic, dynamic> map) => _$GroupSettingsFromJson(map);
 }
@@ -72,12 +89,15 @@ extension BuildFormatExtensions on BuildFormat {
   R fold<R>({
     required R Function() single,
     required R Function() multi,
+    required R Function() tree,
   }) {
     switch (this) {
       case BuildFormat.single:
         return single();
       case BuildFormat.multi:
         return multi();
+      case BuildFormat.tree:
+        return tree();
     }
   }
 }