+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0
+- Fix incorrect `format` mapping for format file
+- Fix generation of fields names on single format file
+- Feat added `multi` format, it generates one class for directory
+- Feat added `prefix_class_name` field in group to customize class name generation in `multi` format
+- Feat added `output_file_name` field in group to customize generated output file name
+BREAKING CHANGES:
+    - Changed defaults value for output file name and class name
+    - Renamed `multi` format to `tree` format
+
 ## 0.0.1
 
 - Initial version.